bugfix: use Config instead of ContainerConfig
The "docker build" will create container to run the RUN/CMD/... command
and commit the container as image.

However, the docker image json content has Config and ContainerConfig.
The ContainerConfig is like the state of container which run the specific
command in the Dockerfile.

For example, the "CMD xxx" command is no-op during building. In order to do
no-op in the container, the "docker build" backend will use

  sh -c "#(nop) CMD [xxx]"

as command which will store in the ContainerConfig. And the "xxx" will
be store in the Config.

Therefore, we should use Config when running container, instead of
ContainerConfig.

Signed-off-by: Wei Fu <fhfuwei@163.com>